Alterntaive to DomainType
This implements a different approach to converting between types and
Datatype dependent type checks.

The main idea here that's different than previously is that the
DomainType enforced a 1:1 correspondence between TileDB's Datatype
enumeration and the Rust type system. Unfortunately, multiple TileDB
Datatypes map to the same Rust type. The idea here is to separate the
conversion of types and the validation that a given Rust type is valid
for a given TileDB Datatype.

The basic idea is to move type conversion between Rust and FFI types
(i.e., i32 -> uint32_t) to be a separate concern from "Is an i32 valid
as a Datatype::Blob". The type conversions basically map directly to how
DomainType worked. The validation is moved to a generic method on
Datatype in tiledb-sys that makes heavy use of `std::any::TypeId` to
restrict the generic type paramter to specific Datatype enum variants.