Fix the VFS::ls and VFS::ls_recursive APIs
These weren't very Rusty on my first attempt. After some discussion with
Ryan, he reminded me that FnMut is a thing and it turns out is actually
usable as an argument being passed to a C API. Its a bit nutty on the
implementation, but it seems to work just fine.