Fix nightly build CI pipeline
I was missing the RELEASES_URL environment variable in the nightly
workflow. Given that I grumbled about the possibility of divergence
while I was first writing this, I took the time to figure out how to
properly reuse steps between workflows. This involved creating a custom
"composite action". Once I figured it out, it turned out to be not too
terrible.