[WIP] Just some work from me refactoring to learn
This is just a quick draft PR to show others some of the refactoring I
did while trying to learn proptest based on what @rroelke has already
written. If we find parts that are good we can extract them for a
separate real PR and eclude all of the stuff I changed on a whim.