Implement Query API steps and adapters, and demonstrate with "reading_incomplete" example (#51)
This branch makes significant adjustments to our query API, focusing principally on the data reading side, and demonstrates those adjustments using the reading_incomplete.rs example which was approximately ported from C.

*Write Queries*
Though not the emphasis of this PR, these did merit some changes which now live inside the query/write module. We add a trait DataProvider which structures can implement so as to manifest data into a query.

*Read Queries*
Previously the query read API required the user to attach their own buffer and manually reallocate it. This is very C-like, and therefore error prone; our Rust layer can offer this as a primitive, but then build adapters on top of it to enable more user-friendly ways to query data.

That's what this pull request does, adding a few categories of query readers:
- the ReadCallbackQuery, which runs a function on top of the query results after each step;
- the TypedReadQuery, which constructs a value from the query results;
- the ManagedReadQuery which allows the user to defer the details of memory allocation to the inside of the query.

There is a lot of glue to tie all of these together, which mostly lives inside query/read/output.rs.
- ScratchAllocator is a trait which can be used to hand objects to ManagedReadQuery
- FixedDataIterator and VarDataIterator are structs to iterate over fixed-length and variable-length data respectively
- FromQueryOutput is a trait for types which can be constructed from an iterator over raw query results

Read queries run in "steps" which run the query until it has written as much as it can to the available buffers. See ReadStepOutput in query/read/mod.rs.

*Example*
The reading_incomplete.rs example is added with three different ways of querying data:
- read_array_step is the closest to the C API, running the query one step at a time and manually reallocating buffers when necessary
- read_array_collect collects array results into Vec<_> structures, and then iterates over those
- read_array_callback registers a callback to be run upon the query results as they arrive