Fix slow property tests
There were two tests that were each taking over twenty seconds to
complete. After some investigation with Ryan the other day, we narrowed
it own to constantly reparsing the pattern passed to `string_regex`. The
idea then was to put all of our `string_regex` patterns behind a
lazy_static. However, it turns out that its none of our strategies
causing the issue.

The issue is `vec(any::<String>(), 0..1024)`. That will constantly
re-parse the default `string_regex` pattern of `\\PC*`. To avoid that
slowness I've created a new strategy that generates `Vec<Vec<u8>>` and
then prop_maps the inner vector into a String. This works because I'm
only generating `u8` values in the range [1, 127) which is guaranteed to
be 7-bit clean and infallibly convertible to UTF-8.