Replace capi_return with capi_call
This removes all access to the underlying raw context pointer except
through the `Context::capi_call` method. This method accepts a single
argument that is a `FnOnce` callable. This callable must take a single
argument which is a `*mut ffi::tiledb_ctx_t` that can be used to pass to
all of the FFI calls that require it. The return value should be the i32
result of invoking one of the tiledb FFI functions.

A simple example looks like such:

```rust
pub fn version(&self) -> TileDBResult<i64> {
    let mut c_version: std::os::raw::c_int = out_ptr!();
    self.capi_call(|ctx| unsafe {
        ffi::tiledb_array_schema_get_allows_dups(
            ctx, c_schema, &mut c_version
        )
    })?;

    Ok(c_version as i64)
}
```