Proc-macros, begone!
Last time I worked on removing the 'ctx lifetime I was too lazy to
figure out how to do it without completely removing everything touched
by the ContextBound and Query derive proc macros. This time I took the
time (literally hours) to undo these so that I canunderstand the lifetime
dances they were doing.

I now have some new thoughts on proc-macros:

First, this is a +394/-859 line change just to *remove* these macros. Which
is to say, it takes less code to implement the thing that the proc macros
were supposed to help with. That seems not great.

Second, some of these changes were non-obvious and surprising in terms
of how they ended up being applied. Having reviewed and +1'ed both of these,
that worries me considerably in terms of adding proc macros. I'm not saying
that we necessarily need to avoid proc macros altogether, but that we should
be a lot more hesitant to reach for them.

So I'm proposing a new rule: proc macros should only be added to
reduce or remove complexity of existing code. If an implementation of
whatever relies on a proc macro, I will now treat that as a code smell
that we're doing something very wrong.