Remove the 'ctx lifetime
In hindsight, this is a shining example of where we should use an
Rc<RefCell<T>> as it exactly matches the semantics of what we want for
Context's. There's a decent bit of churn here, but its all mechanical
and basically just boils down to removing every occurrence of 'ctx in
the code base.