Remove the 'ctx lifetime (#83)
* Wrap RawContext with Rc<RefCell<T>>

This change will allow use to remove all 'ctx lifetime annotations.

* Proc-macros, begone!

Last time I worked on removing the 'ctx lifetime I was too lazy to
figure out how to do it without completely removing everything touched
by the ContextBound and Query derive proc macros. This time I took the
time (literally hours) to undo these so that I canunderstand the lifetime
dances they were doing.

I now have some new thoughts on proc-macros:

First, this is a +394/-859 line change just to *remove* these macros. Which
is to say, it takes less code to implement the thing that the proc macros
were supposed to help with. That seems not great.

Second, some of these changes were non-obvious and surprising in terms
of how they ended up being applied. Having reviewed and +1'ed both of these,
that worries me considerably in terms of adding proc macros. I'm not saying
that we necessarily need to avoid proc macros altogether, but that we should
be a lot more hesitant to reach for them.

So I'm proposing a new rule: proc macros should only be added to
reduce or remove complexity of existing code. If an implementation of
whatever relies on a proc macro, I will now treat that as a code smell
that we're doing something very wrong.

* Remove the 'ctx lifetime

In hindsight, this is a shining example of where we should use an
Rc<RefCell<T>> as it exactly matches the semantics of what we want for
Context's. There's a decent bit of churn here, but its all mechanical
and basically just boils down to removing every occurrence of 'ctx in
the code base.

* Fix two minor bugs while writing proptests

I spent a day today taking another stab at learning the proptest API. So
far my FilterListData Strategy is generating about 100k/sec valid cases.
During that work I found a couple minor bugs in our Rust logic that
contradicted core.

The first bug is a simple typo fixed by doing a `s/Boolean/Blob/`.

The second one is actually interesting. [The C++ code looks like this][1]:

```c++
bool CompressionFilter::accepts_input_datatype(Datatype input_type)  const {
    auto this_filter_type = compressor_to_filter(compressor_);
    if (this_filter_type == FilterType::FILTER_DOUBLE_DELTA ||
        this_filter_type == FilterType::FILTER_DELTA) {
      // Delta filters do not accept floating point types.
      if (datatype_is_real(
        reinterpret_datatype_ != Datatype::ANY ? reinterpret_datatype_ :
                                                 input_type)) {
        return false;
    }
  }

  return true;
}
```

The previous Rust translation of that logic [looked like this][2]:

```rust
FilterData::Compression(CompressionData { kind, .. }) => match kind
{
    CompressionType::Delta {
        reinterpret_datatype,
    }
    | CompressionType::DoubleDelta {
        reinterpret_datatype,
    } => {
        // these filters do not accept floating point
        let check_type =
            if let Some(Datatype::Any) = reinterpret_datatype {
                *input
            } else if let Some(reinterpret_datatype) =
                reinterpret_datatype
            {
                reinterpret_datatype
            } else {
                return None;
            };
        if check_type.is_real_type() {
            None
        } else {
            Some(check_type)
        }
    }
    _ => Some(*input),
},
```

And the corrected translation looks like this:

```rust
FilterData::Compression(CompressionData { kind, .. }) => match kind
{
    CompressionType::Delta {
        reinterpret_datatype,
    }
    | CompressionType::DoubleDelta {
        reinterpret_datatype,
    } => reinterpret_datatype.map_or(Some(*input), |dtype| {
        if !dtype.is_real_type() {
            Some(dtype)
        } else {
            None
        }
    }),
    _ => Some(*input),
},
```

That's fairly subtle.

The reason the first Rust version is wrong is because it misses the one of four
cases where reinterpret_datatype is None and the input type is valid for the
filter which doesn't happen very often given the specificity of the input
datatype.

I love proptets.

[1]: https://github.com/TileDB-Inc/TileDB/blob/d00866b64eab47929170fe61f3498319d7b9711d/tiledb/sm/filter/compression_filter.cc#L98-L111
[2]: https://github.com/TileDB-Inc/tiledb-rs/blob/94cdfa59064548e3e43bc77265c5022324ae73ad/tiledb/api/src/filter/mod.rs#L230-L248

* Remove unnecessary RefCell

* Remove commented code

Turns out this wasn't even necessary to enforce the behavior that I was
wanting. Giving the trait a `Context` return instead of `&Context` means
that we have to call `clone()` when there's a Context member anyway.
Everything else is just passing the value up the chain.

* Don't apply bitwidth reduction against Blobs

* Only put the bitwidth reduction filter first

Rather than fight with reinterpret types, this just only allows it in
the first position.