Update CI to use nightly builds of core (#100)
* Update build system to use core nightly builds

* Fixup API coverage tooling

* Disable ls_recursive test

Apparently this was removed before the release went out so I've just
ignroed the test for now. Theoretically we'll have to test it
eventually.

* Fixup the filter strategy for delta compression

This just restricts the reinterpret_datatype to have a size that evenly
divdes the input datatype so that the buffer length checks aren't
violated in core.