Replace buffer management with Arrow buffers
This implements a second version of the Query interface using Arrow
Arrays. The core idea here is that we can share Arrow arrays freely and
then convert them to mutable buffers as long as there are no external
references to them. This is all done safely and returns an error if any
buffer is externally referenced.