Update PR CI to use new install action
This cleans up the Pull Request CI so that it uses the new install
action correctly. This also removes the weird layer of indirection and
adds the nightly lint step we discussed. The nightly lint is allowed to
fail without breaking the CI. I'm just adding it so we get an early
heads up instead of relying solely on nightly failures.