Create nightly release packages (#199)
* Update Nightly CI

This is some preparatory work before adding the release pipelines. This
is because upstream does not release static builds which is required by
downstream projects that require static builds.

* Remove support for building libtiledb directly

This now relies on having a version of TileDB installed on the system
rather than attempting to build it ourselves.

* Remove old unused CI workflows

These are both replaced by the new packages and nightly-ci workflows.

* Update PR CI to use new install action

This cleans up the Pull Request CI so that it uses the new install
action correctly. This also removes the weird layer of indirection and
adds the nightly lint step we discussed. The nightly lint is allowed to
fail without breaking the CI. I'm just adding it so we get an early
heads up instead of relying solely on nightly failures.

* Update docs on static linkage in README