chore(build): Strip quotes from tiledb prefix (#207)
For some reason, the `pkg-config` that comes in the manylinux containers
does not strip quotes. This avoids using the built-in variables and
instead just grabs the prefix, strips quotes, and manually appends `lib`
for now. Theoretically that could bite me later if something moves to
`lib64` but I'll deal with that if and when it happens.