Complete C++ API Declaration
This is the entirety of the `tiledb/sm/cpp_api` directory in
TileDB-Inc/TileDB. Mostly this is just copied over headers that I then
went through and removed all the fluff and anything that was obviously
not compatible with Rust interop (mostly a bunch of complicated
templates for string vs concrete types).

There are still some unsuported function declarations included which
we'll need to figure out solutions for as they're core APIs that we'll
need to support in some manner. But this should give us the complete
number of patterns that we'll need to translate from C++ to Rust/Cxx
bindings. There are three general groups of C++ functionality that we'll
need to cover:

1. Unwrapped types, i.e., option, pair, unordered_map
2. Templated return types for things like dimension domains
3. Templated arguments for things like query buffers

For 1, I think the existing approach with the current bindings will
suffice. This basically boils down to having a type save bridge and then
just building up the Rust equivalents without using the declared C++
types.

For 2, I'm thinking that we could probably create C++ equivalents to
ScalarValue and Interval and then have some utility functions to convert
back and forth.

Lastly, for 3, I'm planning on running with Arrow Arrays as previously
discussed.

The last important thing to note is that Cxx does not support bridging
templates/generics directly which means we get to define a whole bunch
of functions that are basically `do_thing_u8, do_thing_i16...` as we've
done in a few places in the existing bindings. With luck, we'll be able
to create a few simple macros to reduce the amount of typing required to
implement those each place they're required.